+++ src/js/genghis/json.js
@@ -60,9 +60,20 @@ Genghis.JSON = {
             }
 
             function GenghisDate(date) {
+                function ISODateString(d){
+                    function pad(n){return n < 10 ? '0' + n : n}
+                    d = new Date(d);
+                    return d.getUTCFullYear()+'-'
+                        + pad(d.getUTCMonth()+1)+'-'
+                        + pad(d.getUTCDate())+'T'
+                        + pad(d.getUTCHours())+':'
+                        + pad(d.getUTCMinutes())+':'
+                        + pad(d.getUTCSeconds())+'Z'
+                }
+
                 return {
                     '$genghisType': 'ISODate',
-                    '$value': date ? new Date(date).toString() : null
+                    '$value': date ? ISODateString(date) : null
                 };
             }