+++ CHANGELOG.markdown
@@ -1,6 +1,7 @@
 ## v2.3.1
 
  * [Fix #93] — Skip `bson_ext` and `json` checks for JRuby, et. al.
+ * Use an explicit ISO date string when saving ISODate values.
 
 [i93]: https://github.com/bobthecow/genghis/issues/93